+++ interfaces/Referral-App/src/app/models/sub-category.model.ts
@@ -1,7 +1,0 @@
-export class SubCategory {
-  subCategoryID: number;
-  subCategoryName: string;
-  subCategoryIcon: string;
-  subCategoryDescription?: string;
-  categoryID: number;
-}