+++ interfaces/Referral-App/src/app/pipes/category-filter.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { CategoryFilterPipe } from './category-filter.pipe';
-
-describe('CategoryFilterPipe', () => {
-  it('create an instance', () => {
-    const pipe = new CategoryFilterPipe();
-    expect(pipe).toBeTruthy();
-  });
-});